+++ dantro/plot/funcs/__init__.py
@@ -4,5 +4,5 @@ creators.
 """
 
 from .basic import lineplot
-from .generic import errorbars, facet_grid
+from .generic import errorbars, facet_grid, scatter3d
 from .multiplot import multiplot